Reverse load command mod list on unload mode
When mode is set to unload, 'module load' commands in modulefile are
interpreted as 'module unload' commands. To guaranty correct behavior
regarding requirements, the module list passed to the load command is
reversed to get and unload of the modulefiles in the reverse order than
load.

This is the same behavior than applied on purge or restore commands,
where loaded module list is reversed to be unloaded.

Add non-regression tests in 150-module of 50-cmds suite to check load
and unload of meta-module works correctly over time.