Add indir arg to getPathToModule proc
Be able to optionally precise in what module directory to search for mod
instead of looking for mod in every directory set in MODULEPATH. If
'indir' arg is set, do not raise the 'cannot locate mod' error.

Also build return list with module found status when a modulefile cannot
be returned. Modulefile is then empty string and instead of module name
boolean is set depending if a module has been found for search or not.