Put call to Modules{Display,Help,Test} procs under catch
Put the ModulesDisplay, ModulesHelp and ModulesTest procedures call
under the same catch than used for the evaludation of the modulefile
content. Doing so an error found within these procedures will be handled
the same way than if the error is triggered from the main body of the
modulefile.

As a consequence, the ModulesDisplay, ModulesHelp and ModulesTest
procedures are not anymore called when an error has been raised during
the modulefile main evaluation, since the same catch is shared.

Some existing test results have been adapted for this change and new
tests have been added to check this specific behavior in 070-display,
080-help and 085-test in 70-maint suite.