Add tests for use/unuse of empty path name
Adapt cmdModuleUse to handle empty path name "" as any other path
string, so an error is returned when file type is cheked.

Add tests in 10-use and 50-cmds suites to check behavior of use and
unuse commands with an empty path passed as argument.

Add by the way tests that were missing for bash shell in 10-use suite.