Reuse interp rather creating new one each time
Make use of the same Tcl interp for each modulefile interpreation and
use another one for each modulerc (but the same for each modulerc). By
doing so we proceed like on C-version where same interpreter is used
across modulefile or modulerc interpretation. Huge performance
improvement is achievied with this change on commands making intensive
use of interp like 'avail'.

Change code to use a static name for those interpreter: '__modulefile'
for the interp dedicated to modulefile interpretation and '__modulerc'
for the interp dedicated to the RC interpretation.