Reset interp state from one interp to another
When creating interpreter, dump its variable and procedure content so
before each usage of this interpreter for another modulerc or modulefile
the initial variable and procedure state is restored to avoid spread of
definitions from one interpretation to another.

Variables are removed if they are not part of the initial state or
restored if their value has changed. Procedures are removed if they are
not part of the initial state. Nothing is done if procedures part of the
initial state have been changed during an interpretation as there is no
vital procedure in this list. Interpreter aliases are redefined before
each interpretation to clear any potential override of these aliased
procedures.

dumpInterpState and resetInterpState procedures are introduced to handle
the dump/reset mechanism of the interpreter.

Non regression tests 300-interp in 50-cmds suite have been added to test
the correct behavior of interpreters when running a list of modulerc and
modulefiles.