Normalize collection access issue error message
Test behaviors when collection cannot be accessed with 080-access tests
added to 61-coll suite.

Normalize error messages for the various collection-related commands
when collection cannot be accessed. Catch also error when trying to
access the collection directory but access rights on this directory
does not allow for listing its content.