Distinguish collection found from read test
Previsously cmdModuleRestore and cmdModuleSaveshow procedures were
directly calling for 'file readable' so existence and readability were
tested in a single call. To be able to distinguish an exist issue
from a read issue these two procedures now only test existence and
readability is tested by the readCollectionContent procedure.

As a result, if a collection exists but is not readable a different
message will now appear.