Correct inclusion of aliases in getModules res
Fix inclusion of aliases in getModules procedure results not to include
these aliases if they are not part of the exact same path than path of
the search. Previously if alias source path and search path were
similar, aliases were wrongly included in results.

Add test in 030-multiple of 90-avail to check correct behavior.