Rework default/latest avail version search
Improve getModules procedure to register for each directory the default
element in this dir (implicitly or explicitly set) and the list of all
elements dictionarily sorted.

Then base the code in listModules procedure that look for default or
latest versions only on the information stored in getModules array
result rather than comparing every result names between each other.

As a side effect, default or latest available version printing does not
confuse output anymore when at a given level a sub-directory element is
last element in directory among modulefiles. Previously sub-directory
was printed but last file among modulefiles was also printed (2 latests
version at the same level).

Also to simplify behavior, and understand a directory tagged "default"
does not appear anymore in default listing result as its content (the
default version found in that dir) will be displayed.

Remove stringDictionaryCompare procedure that is not used anymore due to
the code rework.

Add and adapt tests in 020-avail of 91-suite.