Handle empty dir in getModules dir consolidation
When consolidating directory content information in getModules
procedure, handle empty directories so they are removed from the result
array. When an empty directory is removed and its parent directory
becomes empty after child removal, remove also this directory and so on.

Moreover when an empty directory is removed and if it was the default
version of this parent directory, default is changed to last element in
parent directory.

Add and adapt tests in 020-avail of 91-sort suite.