Take benefit of new dir handling in getPathToModule
Rework code in getPathToModule procedure to use the information that is
now stored in the mod_list structure (result of the getModules procedure
call). With this information it is now easier to get directory default
element. Also empty directories are not part of mod_list result, so less
tests are required in getPathToModule.