Register sym versions in getModules result array
Add tracking of the source file that defined a given symbolic module
version in order to add in getModules these symbloc versions to the
'mod_list' result array if they match searched modulepath directory and
module.

Symbolic versions are added to 'mod_list' array as 'version' entries so
they can be ignored by treatment not requiring them (like display
procedures that get symbols not independently but through the relative
modulefile).