Register in getModules dir default even if broken
Register a default set for a given module in the corresponding
'directory' entry in the result consolidated by the getModules
procedure, even if this default does not correspond to a valid entry.
This way it is up to the procedure treating the result provided by
getModules to handle the verification of the default correctness and
display any error.

As a consequence, a bad default will lead to no result displayed for the
corresponding module in case of 'onlydefault' avail display.

Add tests in 020-avail of 91-sort suite to check behavior.