Rework handling of global/user rc defs in getModules
Introduce a 'g_rcVersion' array to handle symbolic versions set from a
global or a user RC the same way than aliases with the 'g_rcAlias'
array.

Then in getModules procedure treat these definitions from global or user
RC with a 'add_rc_defs' flag. If set and if aliases or symbolic versions
match passed mod, these aliases or versions from global or user RC are
added to the result array.

If search in getModules is set to 'rc_defs_included', the aliases and
symbolic versions set from global or user RC added to the result array
in addition to the search result from the passed directory.