Look at loaded mods in getPathToModule if asked only
Add 'look_loaded' argument to getPathToModule procedure in order to look
at the loaded modules to get path for specified 'mod' only if this new
argument is enabled. 'look_loaded' argument is enabled by default.

Disable this behavior when getPathToModule recursively call itself to
solve aliases or when cmdModuleSearch or cmdModulePaths also attempt to
solve alias in the middle of their own search.

Add non-regression tests in 050-locrc of 20-locate suite.