Include global/user rc defs in search/whatis/paths
Ask getModules procedure to include global or user RC definitions when
called in cmdModuleSearch and cmdModulePaths. By doing so aliases or
symbolic versions defined in these global or user RC will be found and
solved if specified as module to search on these commands.

To do so, make 'search' argument of getModules procedure a list rather a
string to be able to precise multiple search keywords like done by
cmdModuleSearch.

Add non-regression tests in 060-rc of 20-locate suite.