Same res for version than alias on getPath/search/paths
Take symbolic version into accounts like done for alias when attempting
to find out the modulefile targeted by a given module name. Symbolic
versions were until now left away from this resolution treatment as
their modulefile target were automatically part of the mod_list result
returned by getModules procedure. But in some cases, a symbolic version
may be hold on an alias, so the target modulefile of this symbol does
not hold the same module name than this symbolic module and thus is not
part of the result list.

This change is useful for resolution procedures like getPathToModule,
cmdModuleSearch and cmdModulePaths. It also helps to localize the
element in error, if any, as the availability of the symbolic version
information will help to determine if the error concerns the search made
by the user or not.

Add relative non-regression tests in 060-rc of 20-locate suite.