Rework registering of access/invalid file error
Integrate modulefile or directory access issue or invalidity directly in
the result array produced by findModules/getModules procedures. By doing
so access to the error information is straigthforward for
getPathToModule procedure.

To achieve this integration, the checkValidModule procedure has been
revised to return any error message in addition to the check status.