Rework error handling in cmdModuleSearch
Take benefit of changes made on getPathToModule to return error
information through the result list to avoid extra call to
getPathToModule in cmdModuleSearch if nothing was found.

Now through the parsing of getModules results and getPathToModule alias
resolution results, all the error information are gathered and
registered if element name matches search string. At the end of the
cmdModuleSearch procedure, all the registered errors are printed.

Add a 'reportIssue' procedure to help reporting error messages depending
of their kind, useful for getPathToModule and cmdModuleSearch
procedures.