Call for getModules in any cases in cmdModuleAliases
Previously getModules was only called when g_moduleAlias and
g_moduleVersion global arrays were both empty, but in case of a global
or user RC file that set an alias or a symbolic version, no more alias
and version are looked for in module paths. To fix that, getModules is
called for all defined module paths in any cases as it is done for
cmdModuleAvail.

Add relative non-regression test in 130-aliases of 70-maint suite.