Relax constraint on cmd line args position
Look the full argument list to find options or command switches, so
options may be passed after the command name and command switches may be
found before the command name. With this behavior commad-line argument
position constraint is aligned with C-version.

Adapt non-regression tests in 70-maint suite and diff doc.