Keep 'module load' lines after init{rm,clear}
Instead of removing the whole line when no more module specified on it
during an 'initrm' command or during an 'initclear', keep the line with
'module load' to be able to reuse this line on next 'initadd' command.

Adapt non-regression tests in 050-rm and 060-2nd of 60-initx suite and
diff doc.