testsuite: refactor tests with code output
Adapt test*out* procedures to call for 'shell_out' procedure if passed
code output is a code order list rather than an output string.

'shell_out' procedure in turn calls for specific procedure to parse
shell commands and transform them in the appropriate shell code. This
way shell code is handled in a central location and tests do not need
anymore to handle the particularity of each supported shells.