Raise and render error once on unload
During an 'unload' action, an error occuring on a loaded module will not
affect the unload of the other modules passed on the command line. This
commit remove the additional error message that were displayed in this
situation: only the message raised from the modulefile interpretation is
useful. Also only render error once, so 'renderError' procedure will do
nothing if it has already been called.