testsuite: fix xres orders production
Use same mechanism than in modulecmd.tcl to get exact same ordering of
x-resource directives to test results.

Enable use of a virtual display server on CentOS test environment with
Xvfb to test behavior of x-resource on this distribution.

Thanks to Gabe Turner for reporting the issues on the x-resource test
that were due to the rewrite of the non-regression testsuite.