Include all kind of symvers target in getModules res
When parsing symbolic version targets in getModules procedures, these
targets were previously included in search result only if they were
found on the filesystem (file or directory). But sometimes, these
targets may be module alias and not a consistent file or directory.

This situation happens when a module-version is set on an alias which in
turns targets a modulefile in a different directory. When you query the
symbolic version in the end-point modulefile directory, you get the
modulefile as a result, but if you query the symbol in the alias
directory, this alias will be returned.

Adapt non-regression tests in 060-rc of 20-locate suite and in
068-transitive of 80-deep suite. Improve especially the 068-transitive
tests with aliases and symbols set from different module.