Fix symbol spread when set on alias before their def
Propagate symbols set on aliases before the definition of these aliases.
Previously, a symbol set on a not-yet-defined alias were lost and not
applied to the alias target once set.

With this change setting a symbol on an alias gives same result either
this symbol is set before or after the alias definition.