Locate module using loaded context only on 'unload'
When searching for a modulefile corresponding to a given name, the
getPathToModule procedure was previously looking by default at the list
of loaded modulefiles to find a match. This behavior lead to different
result depending on the loaded environment context for all module
commands relying on modulefile location.

For instance, if 'mod' had 2 versions '1' and '2', the latter being
default version, if 'mod/1' were loaded it was this version that were
answered when asking for 'mod'.

This commit restore a predictable and idempotent behavior where the
loaded modules context is only use when working on this loaded context.
Same behavior than C-version is now encountered.

As a result, only the 'unload' command now looks at loaded modules to
find location of modules passed as argument. Load, display and help
commands do not look anymore at loaded modules to locate their module
arguments.

Non-regression tests 038-defcontext in 20-locate have been introduced to
ensure correct behavior will persist over time.