Unload deep mod with just root name passed as arg
Enable to unload 'mod/dir/subdir/vers' when unload of 'mod' or 'mod/dir'
asked. Was previously working only if deep module to unload was also the
default version for these root names.

Introduce getLoadedMatchingName procedure to check passed name is equal
to or is a subpart of a loaded module. This new procedure is then used
in getPathToModule when look_loaded is enabled (so when unloading
module).

Add 021-unload tests in 80-deep suite to check unload of deep modules
using sub names. Add also new tests in 180-unload of 50-cmds suite to
check behavior when multiple loaded modules match searched name.