Enable getPathToModule to return closest loaded mod
If 'look_loaded' argument of getPathToModule procedure is set to 'match'
look at matching loaded modules to get result (unload behavior). Enable
to return the closest loaded modules (not module expressed in mod but
sharing at least the same root name) by setting look_loaded argument to
'close'.

With this change, cmdModuleUnload is adapted to call getPathToModule in
'match' look_loaded mode most of the time, but in case of 'module
switch' with a single argument, cmdModuleUnload will call
getPathToModule in 'close' look_loaded mode.

As a consequence, cmdModuleSwitch does not need anymore to work on
module resolution and rely entirely on cmdModuleLoad and cmdModuleUnload
which in turn rely on getPathToModule.

Add complementary tests in 50-cmds/170-swap and 80-deep/022-switch to
ensure closest loaded module is not returned in case of 'switch old new'
command (not 'switch new').