Handle alias/sym pt. to another path in getPathToModule
When searching for a module in a given modulepath directory, if a module
alias or a symbolic version matches searched module but the target of
this alias or symbol is not found in current modulepath directory,
search for this target restarting search from the first modulepath in
list to ensure modulepath priority.

Previously this kind of alias or symbol targeting a module in a
different modulepath were only found if this path was set after the
modulepath where the alias or symbol is defined.

Clarify alias and symbolic version modulepath resolution steps in
modulefile man page.