Add support for hidden 'dot' modulefiles
Modulefiles with part of their name starting with a '.' dot character
are considered as hidden modulefiles, like dot files on Unix-kind
filesystems. After introduction of the getModules/findModules procedures
hidden modulefiles where automatically discarded. So this commit
re-introduce the support to locate these dot modulefiles.

Add a 'fetch_hidden' argument to the findModules procedure to enable an
extra lookup on walked directories to gather the dot modulefiles. The
getModules procedures now looks at the passed mod and if named module
correspond to a dot modulefile findModules 'fetch_hidden' argument is
turned on. Since getPathToModule call to getModules for broad search
(with just passed mod root name), this procedure also makes the
'is hidden module' test to order getModules to do the hidden search.

A hidden modulefile does not appear in case of wild search, it is only
returned when search is about its exact name. When a symbolic version
targets a hidden modulefile, this modulefile will also be returned in
case search is about symbol exact name. A module alias targeting a
hidden modulefile is returned as any other aliases.

Add 067-hidden non-regression tests in 20-locate suite.