Print more readable messages on reportInternalBug
Improve readability of error messages encountered during modulefile
execution by putting Tcl error message first after the 'Module ERROR'
prefix.

To achieve this improvement, modulefile is passed to reportInternalBug
procedure as a separate argument. Procedures getting and manipulating
error information have been adapted to add modulefile as an element of
the error information to be able to transmit it to reportInternalBug.

In addition, this commit change error reporting kind when facing access
error on readModuleContent. Previously access issues were reported with
reportInternalBug whereas these issues were not relative to modulefile
content. reportError is now used to report access issue as when this
kind of error is encountered when checking file is valid modulefile.

Adapt relative non-regression tests and take opportunity to centralize
common error messages at start of the testsuite.