Do not exit if internal error occurs in interp mod
Previously an internal error occurring during the interpretation of a
modulefile (like broken Tcl code) was leading to an immediate exit of
the whole module execution. So the end of the ongoing command was
skipped (like printing separator line '---' on display command or
interpreting next modulefile passed argument on load command).

With this change, an internal error occurring during the interpretation
of a modulefile will be treated as any other error raised by modulefile
Tcl commands (like prereq, conflict, append-path, etc). Interpretation
will be considered as failed but ongoing command will get this failed
status and will continue.

As a result, when this kind of error occurs now display, help, load,
unload, whatis and source commands will end treatment of the failed
modulefile and then proceed with the following modulefiles passed as
arguments.

Add 140-multiargs non-regression tests in 70-maint suite to check
behavior of module commands running with multiple modulefiles passed as
argument.