Only one sep line between 2 mods on display/help/test
Print only one separator line (---) between the interpretation of 2
modules on the display, help and test commands to improve readability of
the results displayed.

To do so, cmdModuleDisplay and cmdModuleTest procedures have been
adapted to receive the full list of modules to work on, like it was
already done on cmdModuleHelp.

Adapt non-regression tests output to this change.