Failed load/unload if submod load/unload fails
When a modulefile loads or unloads another modulefile, load or unload
issue did not previously led to upper module load or unload failure. As
a result, the upper modulefile appeared loaded or unloaded whereas its
dependent sub-modulefile were not loaded or respectively still loaded.

This commit makes cmdModuleLoad and cmdModuleUnload procedure return an
exit code to let know if load or unload action performed correctly. In
case of a 'module load' or 'module unload' call within a modulefile, the
module procedure will raise an error with the dummy message
'SUB_FAILED'. This message is caught at the result of the modulefile
interpretation to return an error code, without raising error count or
message as it has already been handled at the root of the issue.