Handle case when pager is ended (killed or quit)
Catch cases when pager is ended before all messages are reported to it
('q' pressed or process killed).

Error raised from 'puts' call are ignored with 'catch'. As we handle
here informational messages, do not end the full modulecmd process when
message channel is dead.

Also protect pager fd operations in cleanupAndExit, in case pipe is not
operational anymore.