Validate variable names set in MODULES_RUN_QUARANTINE
Test that the names of the environment variable defined in
MODULES_RUN_QUARANTINE are valid names.

Add test in quarantine setup code within init scripts (silent check) and
also test in modulecmd.tcl. The latter reports a warning if a bad name
is detected.

Add a non-regression test in modules.00-init and install.00-init test
suites.