Put autoinit code in braces as much as possible
Use braces for the different shell autoinit code to avoid special
character espacing as much as we can in these complex piece of shell
code.

In some cases, braces cannot be used: when variable value should be
replaced and when a beginning brace does not match a close brace in the
same string).