Make shells return 'true' when rendering ok
Add an expression for each supported shell to say 'true' when code
rendering is perfomed and things are ok. A 'renderOk' procedure is added
following same principle than 'renderError'.

These code additions keep the ability for the rendering to be called
from the shell module function or from a direct execution. To achieve
that instead of using a 'return true' statement, which would had no
meaning outside of a module function, a true statement is expressed.
Most of the shells will return the result from the last evaluated
statement.

For shells not returning result from the last evaluated statement
(python, cmake), a variable is set to true. This variable will be used
in module function to return the status value.

For the sh/csh/fish shells, no code is rendered if something has already
been sent to stdout, as in this case the last lines will return 'true'
by itself.