Make shells return 'false' rather raising exception
Use an expression for each supported shell to say 'false' when error
occurred during modulecmd.tcl process. This was already the behavior for
the sh/csh/fish shells but a change is made to the other lang shells
where previously an exception was raised in case of error.

Moving from a raised exception to a boolean return seem to be more
appropriate for the calling scripts, which will not explode anymore each
time a modulefile is not found for instance.

Boolean return opens the path to a return value check on all supported
shells (was only possible on sh/csh/fish before that).

Now all shells are aligned with a 'true' value returned in case of
success rendering and a 'false' value returned in case of error.