Render true/false with a variable assignment
Set a local variable on shells to return true or false status, like
already done on cmake and python. This way shells treat the return
status in the same way.

For other shells that cmake and python, it does not change behavior as
the variable assignement returns the value assigned, so it still returns
a boolean as last rendered operation.

'real' shells are not touched as they already return in all cases a
boolean status without any specific assignment.

Prior to modulecmd.tcl execution, the variable used for returned status
assignement is initialized to 'true'. This way we are sure the variable
will always be set and with a default 'true' value we may optimize
rendering as no rendering means true.