Align 'true' rendering of sh/csh/fish with other shells
When calling for a 'true' statement at the end of shell code rendering,
also express a true statement for sh/csh/fish as it is the case for
other shells.

True statement will also be produced even if something has already been
sent to stdout channel. To ensure in all cases that shell code will
finish on a true statement if code is correctly generated.