Spool puts stdout command to render after env changes
Supersede 'puts' command to catch content sent to stdout within
modulefile in order to postpone content channel send after rendering on
stdout the relative environment changes required by the modulefile.

By doing so, environment variable set within modulefile will be
available when outputing directly command to the stdout channel.

A 'putsModfileCmd' procedure is added to catch all 'puts' call in a
modulefile interpretation. This procedure analyzes if the puts command
targets the stdout channel. If so, command argument is saved in a
'g_stdoutPuts' global list to be then processed during renderSettings.
If no, the real puts is called (for stderr send for instance).

Add 50-cmds/310-puts non-regression tests to garanty all kind of puts
command are well processed.

Closes #113