Returning text sub-cmds always returns text
In case of error, message is displayed on stderr but commands will still
return text, an empty string. If no result are found, an empty string is
also set as returning value.

Text is returned as result if 'g_return_text' global variable is
defined. It has the priority over error/false/true returns.

This way text can always be expected as the returned value of a module
sub-command returning text in normal cases.

Adapt non-regression test cases for 'path' and 'paths'.