Correlate full path mod to already loaded mod
Check if a passed full path modulefile correspond to an already loaded
module (passed as short name when it was loaded). To detect this full
path modulefile is-loaded, should not be loaded a second time (with a
different name), should be unloaded.

Add tests on load, unload, swap and is-loaded actions.