@prefix@/etc/rc does not depend on MODULESHOME anymore
Avoid dependency to the MODULESHOME environment variable when looking
for the etc/rc global RC file. To define an environment dynamic global RC
file the MODULERCFILE variable may be used. By not relying on an
environment variable we ensure this etc/rc will always be read, if it
exists, whatever the current environment state.

Location of this etc/rc is based on @prefix@ defined at configure time
which is also the basis for the MODULESHOME value, so this change should
not have an impact if MODULESHOME variable were not manipulated
afterward.

This change is useful when using module through commands resetting the
environment like 'sudo' (see RH#1443124, RH#1139165 [1])

module.1 man page is already mentionning @prefix@/etc/rc, no change
needed there.

Adapt testsuite to provide a way to test the use of a @prefix@/etc/rc
file for the whole 'make test' duration, with 'make install-testetcrc'.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1139165