Distinguish cleared var from bare empty path entry
Sometimes a variable is cleared by defining it empty instead of
unsetting it. This is perticulary the case with LOADEDMODULES which is
often set empty in init script to say nothing loaded.

So this situation need to be distinguished from a path variable set to
an empty path entry.

To distinguish both situations, relative reference counter variable is
used to check if an empty path entry is registered.