Move path cmd arg parse in {add,unload}-path
Move path commands (append-path, prepend-path and remove-path) argument
check and parse in internal unload-path and add-path procedures. So it
can benefit for both external and internal calls and gather all checks
in parsePathCommandArgs.

Update all 'unload-path' and 'add-path' calls for new argument order. In
addition it is not needed anymore to pass a separator character to these
2 procedures as it set to the default value through parsePathCommandArgs.

Adapt non-regression tests.