Transparently bufferize err report before pager init
Define a global g_init_errreport flag to know within report or
reportErrorAndExit procedure if error reporting has been initialized or
not. If not yet initialized, call to output message is saved in a buffer
list.

A 'initErrorReport' procedure is added to switch 'g_init_errreport' flag
and output every messages saved in buffer list.

With this change, it is possible to use reportDebug, reportWarning,
reportErrorAndExit, etc procedures transparently without having to know
if error reporting has been enabled or not. The report procedures can
then be used in siteconfig.tcl extension without messing paging output.

Regarding reportDebug, g_init_errreport enables to know if call has to
be saved since if error report is not initialized, debug mode status is
not yet known.