install: reinit module code in version modfile
Adapt version modulefile, which is deployed in versioning installation
mode, to re-initialize module command during its unload, to restore the
command of the previously loaded version.

Goes through a full re-initialization to overcome incompatible
definition of the module command between version 3 and 4.

This full re-initialization cannot be attempted on scripting shells,
like Perl or Python, as function re-definition is not a common practice
on these languages.