Introduce 'module-virtual' modulefile command
Introduce the virtual module concept with the 'module-virtual'
modulefile command. This new command takes a module name as first
argument and a modulefile location as second argument.

A virtual module stands for this module name associated to a modulefile.
The modulefile is the script interpreted when loading or unloading the
virtual module which appears or can be found with its virtual name.

Like a module aliases, virtual modules are expected to be defined within
modulerc files (at whatever rc level). Virtual modules can be targeted
by aliases or symbolic versions. They also compete with aliases and
regular modules for the implicit default version of a module.

Multiple virtual modules may target the same modulefile, which can
distinguish between these various virtual names by use of the
"[module-info name]" command.

A virtual module cannot be loaded using a full path name (merge of its
module path and its module name) as this file does not exist.

Based on a proof of concept made by Bert Wesarg [1].

[1] https://github.com/bertwesarg/modules-tcl/commit/545e9f