Handle ModulesCurrentModulefile in a push/pop fashion
Refactor handling of the ModulesCurrentModulefile public global variable
to get same treatment like ModuleName with a push/pop mechanism.

Then call to the new pushModuleFile and popModuleFile procedures is
gathered in the execute-modulefile and execute-modulerc procedures to
push filename before interpretation and popping it when interpretation
is done.

Change is that ModulesCurrentModulefile is reset after interpretation,
so when this variable is queried in a master interpretation context it
is possible to know if some file is currently being interpreted or not
(value restored to the empty string).