Start pager only if some text has to be printed
Start pager if enabled at first message to print on stderr. No pager is
started anymore when nothing need to be outputted on stderr.

To achieve that pager setup is split on a configuration and enablement
initialization (initPager procedure) and startup (startPager procedure).
Split in two procedures is required to report debug messages on pager
configuration that will be printed through the started pager.

Partially fix #146.